Fix post-SSH tunnel code. Add option to keep DNS tunnel open.
The original SSH tunneling code displayed incorrect messages after exiting
out of the tunnel. Specifically, as the only way to exit was ^C, it would
always display an error. A better post-SSH tunnel set of messages was
added. The default behaviour when exiting out of the SSH tunnel is to close
the DNS tunnel as well. A command line flag was added to allow keeping the
tunnel open after the SSH tunnel.